Introduce file_regression fixture
