+++ tests/test_file_simple/test_simple_bin_file.bin
@@ -1,0 +1,1 @@
+binary contents ÿÿŞ
